Version bump: 0.9.33  - Implemented new bulk format identification options  - Removed internal copies of fido and droid    From now on this gem has an external dependency on installations and proper configuration of these tools    The upside is that with both these changes we now get a consistent and improved performance for format identification

+++ lib/libis/format/droid.rb
@@ -3,10 +3,6 @@ require 'singleton'
 require 'tempfile'
 require 'csv'
 
-require 'libis/tools/extend/string'
-require 'libis/tools/logger'
-require 'libis/tools/command'
-
 require 'libis/format/config'
 
 unless CSV::HeaderConverters.has_key?(:droid_headers)
@@ -23,11 +19,8 @@ module Libis
 
     class Droid < Libis::Format::IdentificationTool
 
-      protected
-
       def run_list(filelist)
-        output = runner(filelist)
-        process_output(output)
+        runner(filelist)
       end
 
       def run_dir(dir, recursive = true)
@@ -39,12 +32,10 @@ module Libis
       end
 
       def run(file)
-        profile = profile_file_name
-        report = result_file_name
-        create_profile(file, profile)
-        create_report(profile, report)
-        parse_report(report)
+        runner(file)
       end
+
+      protected
 
       def runner(file_or_list)
         profile = profile_file_name
@@ -55,24 +46,28 @@ module Libis
       end
 
       def parse_report(report)
-        keys = [:id, :parent_id, :uri, :filepath, :filename, :matchtype, :status, :filesize, :type, :extension, :mod_time, :ext_mismatch, :hash, :format_count, :puid, :format_name, :format_version]
-        result = CSV.parse(File.readlines(report).join("\n")).map {|a| Hash[keys.zip(a)]}
-        result = CSV.read(report, headers: true, header_converters: [:symbol])
-                     .map {|a| Hash[keys.zip(a.values)]}
+        keys = [
+            :id, :parent_id, :uri, :filepath, :filename, :matchtype, :status, :filesize, :type, :extension,
+            :mod_time, :ext_mismatch, :hash, :format_count, :puid, :mimetype, :format_name, :format_version]
+        result = CSV.parse(File.readlines(report).join)
+                     .map {|a| Hash[keys.zip(a)]}
                      .select {|a| a[:type] == 'File'}
-        File.delete report
+        # File.delete report
         result.each do |r|
           r.delete(:id)
           r.delete(:parent_id)
           r.delete(:uri)
           r.delete(:filename)
-          r.delete()
-          r.delete()
-          r.delete()
-          r.delete()
+          r.delete(:status)
+          r.delete(:filesize)
+          r.delete(:type)
+          r.delete(:extension)
+          r.delete(:mod_time)
+          r.delete(:hash)
           r.delete(:format_count)
           r[:source] = :droid
         end
+        process_output(result)
       end
 
       def create_report(profile, report)
@@ -82,17 +77,18 @@ module Libis
             '-q'
         ]
         result = Libis::Tools::Command.run(Libis::Format::Config[:droid_path], *args)
-
-        warn "DROID report errors: #{result[:err].join("\n")}" unless result[:status] == 0
+        raise RuntimeError, "DROID report errors: #{result[:err].join("\n")}" unless result[:status] == 0
         File.delete profile
       end
 
       def create_profile(file_or_list, profile, recursive = false)
-        args = (file_or_list.is_a?(Array)) ? file_or_list.map(&escape_for_string) : file_or_list.escape_for_string
+        args = []
+        files = (file_or_list.is_a?(Array)) ? file_or_list.map(&:escape_for_string) : [file_or_list.escape_for_string]
+        files.each { |file| args << '-a' << file}
         args << '-p' << profile << '-q'
         args << '-R' if recursive
         result = Libis::Tools::Command.run(Libis::Format::Config[:droid_path], *args)
-        warn "DROID profile errors: #{result[:err].join("\n")}" unless result[:status] == 0
+        raise RuntimeError, "DROID profile errors: #{result[:err].join("\n")}" unless result[:status] == 0
       end
 
       def profile_file_name
